Move validation into its own module
